Split syscalls into separate files
Without this, compilation on Windows doesn't work at all
because the syscalls are platform-specific.
