+++ env.go
@@ -2,16 +2,7 @@ package tableflip
 
 import (
 	"os"
-	"syscall"
 )
-
-var stdEnv = &env{
-	newProc:     newOSProcess,
-	newFile:     os.NewFile,
-	environ:     os.Environ,
-	getenv:      os.Getenv,
-	closeOnExec: syscall.CloseOnExec,
-}
 
 type env struct {
 	newProc     func(string, []string, []*os.File, []string) (process, error)
