+++ env_syscalls.go
@@ -1,0 +1,16 @@
+// +build !windows
+
+package tableflip
+
+import (
+	"os"
+	"syscall"
+)
+
+var stdEnv = &env{
+	newProc:     newOSProcess,
+	newFile:     os.NewFile,
+	environ:     os.Environ,
+	getenv:      os.Getenv,
+	closeOnExec: syscall.CloseOnExec,
+}
