+++ fds.go
@@ -340,12 +340,3 @@ func dupConn(conn syscall.Conn, name fileName) (*file, error) {
 	}
 	return dup, duperr
 }
-
-func dupFd(fd uintptr, name fileName) (*file, error) {
-	dupfd, _, errno := syscall.Syscall(syscall.SYS_FCNTL, fd, syscall.F_DUPFD_CLOEXEC, 0)
-	if errno != 0 {
-		return nil, errors.Wrap(errno, "can't dup fd using fcntl")
-	}
-
-	return newFile(dupfd, name), nil
-}
